feat: 네비게이션 바와 기존 페이지 레이아웃 조정정
