+++ src/pages/LoginPage.jsx
@@ -1,36 +1,39 @@
 /** @jsxImportSource @emotion/react */
 import React from 'react';
 import styled from '@emotion/styled';
+import NavBar from '../components_yjin/NavBar';
 
 function LoginPage() {
   return (
-    <PageContainer>
+    <Wrapper>
       {/* 이미 제작된 헤더를 사용한다고 하셨으므로, 필요한 위치에 삽입 */}
-      {/* <Header /> */}
+      <NavBar />
 
-      <LoginTitle>로그인</LoginTitle>
-      <SubText>자율 스터디 서비스에 오신 걸 환영해요!</SubText>
+      <LoginContainer>
+        <LoginTitle>로그인</LoginTitle>
+        <SubText>자율 스터디 서비스에 오신 걸 환영해요!</SubText>
 
-      <FormContainer>
-        {/* 학번 입력 */}
-        <InputContainer>
-          <StyledInput type="text" placeholder="학번" />
-        </InputContainer>
+        <FormContainer>
+          {/* 학번 입력 */}
+          <InputContainer>
+            <StyledInput type="text" placeholder="학번" />
+          </InputContainer>
 
-        {/* 비밀번호 입력 */}
-        <InputContainer style={{ marginTop: '20px' }}>
-          <StyledInput type="password" placeholder="비밀번호" />
-        </InputContainer>
+          {/* 비밀번호 입력 */}
+          <InputContainer style={{ marginTop: '20px' }}>
+            <StyledInput type="password" placeholder="비밀번호" />
+          </InputContainer>
 
-        {/* 로그인 버튼 */}
-        <LoginButton>로그인</LoginButton>
+          {/* 로그인 버튼 */}
+          <LoginButton>로그인</LoginButton>
 
-        {/* 회원가입 버튼 (텍스트 링크) - 우측 정렬 */}
-        <SignupButtonContainer>
-          <SignupButton>회원가입</SignupButton>
-        </SignupButtonContainer>
-      </FormContainer>
-    </PageContainer>
+          {/* 회원가입 버튼 (텍스트 링크) - 우측 정렬 */}
+          <SignupButtonContainer>
+            <SignupButton>회원가입</SignupButton>
+          </SignupButtonContainer>
+        </FormContainer>
+      </LoginContainer>
+    </Wrapper>
   );
 }
 
@@ -38,19 +41,19 @@ export default LoginPage;
 
 /* ----- 스타일 정의 ----- */
 
-// 페이지 전체 컨테이너 (가운데 정렬 + 고정 너비)
-const PageContainer = styled.div`
-  width: 380px;
-  margin: 0 auto;
+const Wrapper = styled.div`
+  width: 420px;
+  height: 100vh;
   padding: 0 20px;
   box-sizing: border-box;
 `;
 
-// "로그인" 타이틀
-const LoginTitle = styled.h1`
-  margin-top: 72px;
-  margin-bottom: 10px; /* 다음 텍스트와 10px 간격 */
+const LoginContainer = styled.div`
+  padding-top: 152px; /* NavBar와 LoginTitle 사이의 간격을 여기서 조정 */
+`;
 
+const LoginTitle = styled.h1`
+  margin-bottom: 10px; /* 다음 텍스트와 10px 간격 */
   color: #000;
   font-family: Pretendard, sans-serif;
   font-size: 24px;
@@ -60,10 +63,8 @@ const LoginTitle = styled.h1`
   text-align: left;
 `;
 
-// "자율 스터디 서비스에 오신 걸 환영해요!"
 const SubText = styled.p`
   margin-bottom: 72px;
-
   color: #6b6b6b;
   font-family: Pretendard, sans-serif;
   font-size: 16px;
@@ -73,14 +74,12 @@ const SubText = styled.p`
   text-align: left;
 `;
 
-// 폼 영역(전체 폭 380px)
 const FormContainer = styled.div`
   width: 100%;
   display: flex;
   flex-direction: column;
 `;
 
-// 학번/비밀번호 입력 컨테이너 (가로 크기 380px)
 const InputContainer = styled.div`
   width: 100%;
   height: 48px;
@@ -89,24 +88,20 @@ const InputContainer = styled.div`
   gap: 10px;
   padding: 0 12px; /* 좌우 여유 공간 */
   box-sizing: border-box;
-
   border-radius: 8px;
   border: 1px solid #6b6b6b; /* var(--Component-Default-Sub) */
   background: #fff;
 `;
 
-// 실제 입력 필드
 const StyledInput = styled.input`
   flex: 1; /* 남은 공간을 채우도록 */
   border: none;
   outline: none;
-
   font-family: Pretendard, sans-serif;
   font-size: 14px;
   color: #000;
 `;
 
-// 로그인 버튼 (가로 크기 380px)
 const LoginButton = styled.button`
   width: 100%;
   height: 48px;
@@ -129,32 +124,25 @@ const LoginButton = styled.button`
   letter-spacing: -0.5px;
 `;
 
-// 회원가입 버튼을 우측 정렬하기 위한 래퍼
 const SignupButtonContainer = styled.div`
   width: 100%;
   margin-top: 20px;
-
   display: flex;
   justify-content: flex-end;
 `;
 
-// 회원가입 텍스트 버튼
 const SignupButton = styled.button`
   background: none;
   border: none;
   cursor: pointer;
-
-  /* 기본 패딩/마진 제거 */
   padding: 0;
   margin: 0;
-
   color: #000;
   font-family: Pretendard, sans-serif;
   font-size: 16px;
   font-weight: 500;
   line-height: normal;
   letter-spacing: -0.4px;
-
   text-decoration-line: underline;
   text-decoration-style: solid;
   text-decoration-skip-ink: none;
