+++ src/router.jsx
@@ -4,10 +4,11 @@ import Home1 from './pages_yjin/Home1';
 import Notfound from './pages_yjin/Notfound';
 
 // 로그인 페이지 컴포넌트
-import LoginPage from './components/LoginPage'; // 경로에 맞춰 수정
-import SignupPage from './components/SignupPage';
-import StudyApplyeeStatusPage from './components/StudyApplyeeStatusPage';
-import StudyApplierStatusPage from './components/StudyApplierStatusPage';
+import LoginPage from './pages/LoginPage'; // 경로에 맞춰 수정
+import SignupPage from './pages/SignupPage';
+import StudyApplyeeStatusPage from './pages/StudyApplyeeStatusPage';
+import StudyApplierStatusPage from './pages/StudyApplierStatusPage';
+import StudyPage from './pages/StudyPage';
 
 const Router = () => {
   return (
@@ -30,6 +31,9 @@ const Router = () => {
 
         {/* 지원상태확인 페이지 추가 */}
         <Route path="/applierstatus" element={<StudyApplierStatusPage />} />
+
+        {/* 스터디 페이지 */}
+        <Route path="/study" element={<StudyPage />} />
 
         {/* 위에 해당되지 않는 모든 경로 -> Notfound */}
         <Route path="*" element={<Notfound />} />
